requirements.txt
# Project dependencies 
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
django-environ>=0.11.2
djangorestframework-simplejwt>=5.3.0